[Update] SDK version increased to 30
The target SDK version has been changed from 29 to 30 due to API level requirements for Google Play apps.
Also, the distributionSha256Sum was removed because of the following:
WARNING: It is not fully supported to define distributionSha256Sum in gradle-wrapper.properties.
Using an incorrect value may freeze or crash Android Studio.
Please manually verify or remove this property from all of included projects if applicable.