Update GPX export path to Documents folder for user accessibility (#458)
* SDK and MinSKD update

* Change GPX export to user-accessible location

Modifies the GPX export path to save files in a user-accessible location. Previously, files were exported to a default location that was not visible to the user, making access difficult. Exports are now saved in the Downloads folder. The solution works up to the latest version, Android 14.

Co-Authored-By: Kevin Salazar <80077213+Kevin-Salazar-itcr@users.noreply.github.com>

* Update strings-preferences.xml

Changed the GPX export location to the Documents folder for better organization and user accessibility, as per feedback. Adjusted related strings and resolved conflicts with PR #451

* Update GPX export logic

* Update GpsStatusRecord.java

* Update About.java

* Update ExportToStorageTask.java

* Update strings-preferences.xml

* Update ExportToStorageTask.java

* Update ExportToStorageTask.java

* Update AndroidManifest.xml

---------

Co-authored-by: Kevin Salazar <80077213+Kevin-Salazar-itcr@users.noreply.github.com>
Co-authored-by: Binnette <binnette@gmail.com>