Solving issue #347
The problem reported on the issue was wrong, all works good, the text doesn´t seen well, but it has a scrollbar to solve the problem, but when i was testing that i saw that the component on the trackmanager were loking bad and this in the available layouts section too, so i fixed that.

First, to the trackmaneger i put a line to the date or name to see text in a uniform way and separate the waypoints and trackpoints counter in two lines, this with the objetive of a friendly interface.

Second, to the available layouts, i put a scroll bar and the same adjust to the text that i used in the track manager, to see text in a properly way.